requirements.txt
openai~=1.14.2
langchain-community
python-dotenv~=1.0.1
pandas~=2.2.1
fastapi~=0.110.0
langchain~=0.1.13
pydantic~=2.6.4
chromadb==0.4.22
langchain_experimental
tabulate
langsmith~=0.1.31